新增RWKV-7模型及相关音频处理功能，支持音频特征提取和推理。新增ASR推理脚本，集成Whisper特征提取器和编码器，优化了音频模型的加载与处理流程，同时引入了CUDA加速以提升性能。

+++ model/llm/cuda/rwkv7_state_fwd_fp16.cpp
@@ -1,8 +1,8 @@
 #include <torch/extension.h>
 #include "ATen/ATen.h"
 
-// typedef at::Half dtype;
-typedef at::BFloat16 dtype;
+typedef at::Half dtype;
+// typedef at::BFloat16 dtype;
 void cuda_forward(int B, int T, int C, int H, float *state, dtype *r, dtype *w, dtype *k, dtype *v, dtype *a, dtype *b, dtype *y);
 
 void forward(int64_t B, int64_t T, int64_t C, int64_t H, torch::Tensor &state, torch::Tensor &r, torch::Tensor &w, torch::Tensor &k, torch::Tensor &v, torch::Tensor &a, torch::Tensor &b, torch::Tensor &y) {
